@timvandermeij 2019-10-12
Thank you for improving this, and the performance patches in general!